+++ lib/src/main/java/com/orange/ods/compose/component/bottomnavigation/OdsBottomNavigation.kt
@@ -10,7 +10,6 @@
 
 package com.orange.ods.compose.component.bottomnavigation
 
-import android.content.res.Configuration
 import androidx.annotation.DrawableRes
 import androidx.compose.foundation.layout.RowScope
 import androidx.compose.material.BottomNavigation
@@ -18,14 +17,16 @@ import androidx.compose.material.BottomNavigationItem
 import androidx.compose.material.Icon
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.tooling.preview.Preview
 import com.orange.ods.compose.component.OdsComponentApi
 import com.orange.ods.compose.component.utilities.Preview
+import com.orange.ods.compose.component.utilities.UiModePreviews
 import com.orange.ods.compose.theme.OdsTheme
 
 /**
@@ -116,6 +117,7 @@ fun RowScope.OdsBottomNavigationItem(
     )
 }
 
+@UiModePreviews.Default
 @Composable
 private fun PreviewOdsBottomNavigation() = Preview {
     data class Item(@DrawableRes val iconResId: Int, val label: String)
@@ -127,27 +129,15 @@ private fun PreviewOdsBottomNavigation() = Preview {
         Item(android.R.drawable.ic_dialog_info, "Fourth item")
     )
 
-    val selectedItemIndex = remember { mutableStateOf(0) }
+    var selectedItemIndex by remember { mutableStateOf(0) }
     OdsBottomNavigation {
         items.forEachIndexed { index, item ->
             OdsBottomNavigationItem(
                 icon = { Icon(painter = painterResource(id = item.iconResId), contentDescription = null) },
                 label = item.label,
-                selected = selectedItemIndex.value == index,
-                onClick = { selectedItemIndex.value = index }
+                selected = selectedItemIndex == index,
+                onClick = { selectedItemIndex = index }
             )
         }
     }
 }
-
-@Preview(name = "OdsBottomNavigation - Light")
-@Composable
-private fun PreviewOdsBottomNavigationLight() = PreviewOdsBottomNavigation()
-
-@Preview(
-    name = "OdsBottomNavigation - Dark",
-    uiMode = Configuration.UI_MODE_NIGHT_YES,
-    showBackground = true
-)
-@Composable
-private fun PreviewOdsBottomNavigationDark() = PreviewOdsBottomNavigation()