+++ lib/src/main/java/com/orange/ods/compose/component/button/OdsIconToggleButton.kt
@@ -10,7 +10,6 @@
 
 package com.orange.ods.compose.component.button
 
-import android.content.res.Configuration
 import androidx.compose.animation.animateColorAsState
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.foundation.background
@@ -23,14 +22,15 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.orange.ods.compose.component.OdsComponentApi
 import com.orange.ods.compose.component.utilities.DisabledInteractionSource
 import com.orange.ods.compose.component.utilities.Preview
+import com.orange.ods.compose.component.utilities.UiModePreviews
 import com.orange.ods.compose.theme.OdsDisplaySurface
 import com.orange.ods.compose.theme.OdsTheme
 import com.orange.ods.theme.OdsColors
@@ -105,25 +105,14 @@ private fun OdsColors.buttonToggleBackgroundColor(displaySurface: OdsDisplaySurf
         OdsDisplaySurface.Light -> OdsTheme.lightThemeColors.primary
     }
 
+@UiModePreviews.Default
 @Composable
 private fun PreviewOdsIconToggleButton() = Preview {
-    val checked = remember { mutableStateOf(false) }
+    var checked by remember { mutableStateOf(false) }
     OdsIconToggleButton(
-        checked = checked.value,
-        onCheckedChange = { checked.value = it },
+        checked = checked,
+        onCheckedChange = { checked = it },
         icon = painterResource(id = android.R.drawable.ic_btn_speak_now),
         contentDescription = "Microphone"
     )
 }
-
-@Preview(name = "OdsIconToggleButton - Light")
-@Composable
-private fun PreviewOdsIconToggleButtonLight() = PreviewOdsIconToggleButton()
-
-@Preview(
-    name = "OdsIconToggleButton - Dark",
-    uiMode = Configuration.UI_MODE_NIGHT_YES,
-    showBackground = true
-)
-@Composable
-private fun PreviewOdsIconToggleButtonDark() = PreviewOdsIconToggleButton()