+++ lib/src/main/java/com/orange/ods/compose/component/card/OdsSmallCard.kt
@@ -10,7 +10,6 @@
 
 package com.orange.ods.compose.component.card
 
-import android.content.res.Configuration
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
@@ -28,10 +27,10 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.semantics.semantics
-import androidx.compose.ui.tooling.preview.Preview
 import com.orange.ods.R
 import com.orange.ods.compose.component.OdsComponentApi
 import com.orange.ods.compose.component.utilities.Preview
+import com.orange.ods.compose.component.utilities.UiModePreviews
 import com.orange.ods.compose.theme.OdsTheme
 
 /**
@@ -101,6 +100,7 @@ fun OdsSmallCard(
     }
 }
 
+@UiModePreviews.Default
 @Composable
 private fun PreviewOdsSmallCard() = Preview {
     OdsSmallCard(
@@ -109,15 +109,3 @@ private fun PreviewOdsSmallCard() = Preview {
         image = painterResource(id = R.drawable.placeholder)
     )
 }
-
-@Preview(name = "OdsSmallCard - Light")
-@Composable
-private fun PreviewOdsSmallCardLight() = PreviewOdsSmallCard()
-
-@Preview(
-    name = "OdsSmallCard - Dark",
-    uiMode = Configuration.UI_MODE_NIGHT_YES,
-    showBackground = true
-)
-@Composable
-private fun PreviewOdsSmallCardDark() = PreviewOdsSmallCard()