+++ lib/src/main/java/com/orange/ods/compose/component/chip/OdsChip.kt
@@ -10,7 +10,6 @@
 
 package com.orange.ods.compose.component.chip
 
-import android.content.res.Configuration
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.size
@@ -22,8 +21,10 @@ import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.material.Icon
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.res.dimensionResource
@@ -31,7 +32,6 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.semantics.stateDescription
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
 import com.orange.ods.R
@@ -40,6 +40,7 @@ import com.orange.ods.compose.component.chip.OdsChipDefaults.SurfaceOverlayOpaci
 import com.orange.ods.compose.component.utilities.BasicPreviewParameterProvider
 import com.orange.ods.compose.component.utilities.OdsImageCircleShape
 import com.orange.ods.compose.component.utilities.Preview
+import com.orange.ods.compose.component.utilities.UiModePreviews
 import com.orange.ods.compose.theme.OdsTheme
 import com.orange.ods.utilities.extension.noRippleClickable
 
@@ -154,36 +155,17 @@ private fun odsChipBorderColor(selected: Boolean, enabled: Boolean) = when {
     else -> OdsTheme.colors.onSurface.copy(alpha = SurfaceOverlayOpacity)
 }
 
+@UiModePreviews.Chip
 @Composable
-private fun PreviewOdsChip(outlined: Boolean) = Preview {
-    val selected = remember { mutableStateOf(false) }
+private fun PreviewOdsChip(@PreviewParameter(OdsChipPreviewParameterProvider::class) outlined: Boolean) = Preview {
+    var selected by remember { mutableStateOf(false) }
     OdsChip(
         text = "Text",
-        selected = selected.value,
-        onClick = { selected.value = !selected.value },
+        selected = selected,
+        onClick = { selected = !selected },
         leadingAvatar = painterResource(id = R.drawable.placeholder_small),
         outlined = outlined
     )
-}
-
-@Preview(
-    name = "OdsChip - Light",
-    widthDp = 180
-)
-@Composable
-private fun PreviewOdsChipLight(@PreviewParameter(OdsChipPreviewParameterProvider::class) outlined: Boolean) {
-    PreviewOdsChip(outlined)
-}
-
-@Preview(
-    name = "OdsChip - Dark",
-    uiMode = Configuration.UI_MODE_NIGHT_YES,
-    showBackground = true,
-    widthDp = 180
-)
-@Composable
-private fun PreviewOdsChipDark(@PreviewParameter(OdsChipPreviewParameterProvider::class) outlined: Boolean) {
-    PreviewOdsChip(outlined)
 }
 
 internal class OdsChipPreviewParameterProvider : BasicPreviewParameterProvider<Boolean>(false, true)