+++ lib/src/main/java/com/orange/ods/compose/component/chip/OdsChoiceChipsFlowRow.kt
@@ -15,12 +15,18 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.selection.selectableGroup
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.PreviewParameter
 import com.google.accompanist.flowlayout.FlowRow
 import com.orange.ods.R
 import com.orange.ods.compose.component.OdsComponentApi
+import com.orange.ods.compose.component.utilities.BasicPreviewParameterProvider
+import com.orange.ods.compose.component.utilities.Preview
+import com.orange.ods.compose.component.utilities.UiModePreviews
 
 /**
  * Displays a full width [FlowRow] containing customized choice chips [OdsChoiceChipsFlowRowScope.OdsChoiceChip].
@@ -80,3 +86,30 @@ fun <T> OdsChoiceChipsFlowRowScope<T>.OdsChoiceChip(@StringRes textRes: Int, val
  * Scope for the children of [OdsChoiceChipsFlowRow].
  */
 data class OdsChoiceChipsFlowRowScope<T>(val selectedChip: MutableState<T>, val outlinedChips: Boolean)
+
+@UiModePreviews.Default
+@Composable
+private fun PreviewOdsChoiceChipsFlowRow(@PreviewParameter(OdsChoiceChipsFlowRowPreviewParameterProvider::class) outlinedChips: Boolean) = Preview {
+    data class ChoiceChip(val text: String, val enabled: Boolean, val value: Int)
+
+    val choiceChips = listOf(
+        ChoiceChip("First", true, 1),
+        ChoiceChip("Second", true, 2),
+        ChoiceChip("Third", false, 3),
+        ChoiceChip("Fourth", true, 4)
+    )
+
+    val selectedChip = remember { mutableStateOf(choiceChips.first().value) }
+    OdsChoiceChipsFlowRow(
+        selectedChip = selectedChip,
+        outlinedChips = outlinedChips
+    ) {
+        choiceChips.forEach { choiceChip ->
+            with(choiceChip) {
+                OdsChoiceChip(text = text, value = value, enabled = enabled)
+            }
+        }
+    }
+}
+
+internal class OdsChoiceChipsFlowRowPreviewParameterProvider : BasicPreviewParameterProvider<Boolean>(false, true)