+++ lib/src/main/java/com/orange/ods/compose/component/chip/OdsFilterChip.kt
@@ -10,7 +10,6 @@
 
 package com.orange.ods.compose.component.chip
 
-import android.content.res.Configuration
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Box
@@ -23,8 +22,10 @@ import androidx.compose.material.LocalContentAlpha
 import androidx.compose.material.LocalContentColor
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -33,7 +34,6 @@ import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
 import com.orange.ods.R
@@ -42,6 +42,7 @@ import com.orange.ods.compose.component.utilities.BasicPreviewParameterProvider
 import com.orange.ods.compose.component.utilities.DisabledInteractionSource
 import com.orange.ods.compose.component.utilities.OdsImageCircleShape
 import com.orange.ods.compose.component.utilities.Preview
+import com.orange.ods.compose.component.utilities.UiModePreviews
 import com.orange.ods.compose.theme.OdsTheme
 
 /**
@@ -131,36 +132,17 @@ private fun OdsChipSelectedIcon(tint: Color = LocalContentColor.current.copy(alp
     )
 }
 
+@UiModePreviews.Chip
 @Composable
-private fun PreviewOdsFilterChip(outlined: Boolean) = Preview {
-    val selected = remember { mutableStateOf(false) }
+private fun PreviewOdsFilterChip(@PreviewParameter(OdsFilterChipPreviewParameterProvider::class) outlined: Boolean) = Preview {
+    var selected by remember { mutableStateOf(false) }
     OdsFilterChip(
         text = "Text",
-        selected = selected.value,
-        onClick = { selected.value = !selected.value },
+        selected = selected,
+        onClick = { selected = !selected },
         leadingAvatar = painterResource(id = R.drawable.ic_check),
         outlined = outlined
     )
-}
-
-@Preview(
-    name = "OdsFilterChip - Light",
-    widthDp = 180
-)
-@Composable
-private fun PreviewOdsFilterChipLight(@PreviewParameter(OdsFilterChipPreviewParameterProvider::class) outlined: Boolean) {
-    PreviewOdsFilterChip(outlined)
-}
-
-@Preview(
-    name = "OdsFilterChip - Dark",
-    uiMode = Configuration.UI_MODE_NIGHT_YES,
-    showBackground = true,
-    widthDp = 180
-)
-@Composable
-private fun PreviewOdsFilterChipDark(@PreviewParameter(OdsFilterChipPreviewParameterProvider::class) outlined: Boolean) {
-    PreviewOdsFilterChip(outlined)
 }
 
 internal class OdsFilterChipPreviewParameterProvider : BasicPreviewParameterProvider<Boolean>(false, true)