+++ lib/src/main/java/com/orange/ods/compose/component/control/OdsCheckbox.kt
@@ -10,20 +10,21 @@
 
 package com.orange.ods.compose.component.control
 
-import android.content.res.Configuration
 import androidx.compose.material.Checkbox
 import androidx.compose.material.CheckboxDefaults
 import androidx.compose.material.ContentAlpha
 import androidx.compose.material.ripple.LocalRippleTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.tooling.preview.Preview
 import com.orange.ods.compose.component.OdsComponentApi
 import com.orange.ods.compose.component.utilities.Preview
+import com.orange.ods.compose.component.utilities.UiModePreviews
 import com.orange.ods.compose.theme.OdsPrimaryRippleTheme
 import com.orange.ods.compose.theme.OdsTheme
 
@@ -77,24 +78,12 @@ private object OdsCheckboxDefault {
 
 }
 
-
+@UiModePreviews.Default
 @Composable
 private fun PreviewOdsCheckbox() = Preview {
-    val checked = remember { mutableStateOf(false) }
+    var checked by remember { mutableStateOf(false) }
     OdsCheckbox(
-        checked = checked.value,
-        onCheckedChange = { checked.value = it }
+        checked = checked,
+        onCheckedChange = { checked = it }
     )
 }
-
-@Preview(name = "OdsCheckbox - Light")
-@Composable
-private fun PreviewOdsCheckboxLight() = PreviewOdsCheckbox()
-
-@Preview(
-    name = "OdsCheckbox - Dark",
-    uiMode = Configuration.UI_MODE_NIGHT_YES,
-    showBackground = true
-)
-@Composable
-private fun PreviewOdsCheckboxDark() = PreviewOdsCheckbox()