+++ lib/src/main/java/com/orange/ods/compose/component/control/OdsSlider.kt
@@ -10,7 +10,6 @@
 
 package com.orange.ods.compose.component.control
 
-import android.content.res.Configuration
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.BoxWithConstraints
@@ -26,8 +25,10 @@ import androidx.compose.material.SliderDefaults
 import androidx.compose.material.Text
 import androidx.compose.material.contentColorFor
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -35,12 +36,12 @@ import androidx.compose.ui.graphics.compositeOver
 import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import com.orange.ods.R
 import com.orange.ods.compose.component.OdsComponentApi
 import com.orange.ods.compose.component.utilities.Preview
+import com.orange.ods.compose.component.utilities.UiModePreviews
 import com.orange.ods.compose.theme.OdsTheme
 
 private const val ActiveTickColorAlpha = 0.4f
@@ -265,7 +266,7 @@ private object OdsSliderDefaults {
 
 }
 
-
+@UiModePreviews.Default
 @Composable
 private fun PreviewOdsSlider() = Preview {
     val sliderValue = remember { mutableStateOf(0.5f) }
@@ -276,36 +277,13 @@ private fun PreviewOdsSlider() = Preview {
     )
 }
 
-@Preview(name = "OdsSlider - Light")
-@Composable
-private fun PreviewOdsSliderLight() = PreviewOdsSlider()
-
-@Preview(
-    name = "OdsSlider - Dark",
-    uiMode = Configuration.UI_MODE_NIGHT_YES,
-    showBackground = true
-)
-@Composable
-private fun PreviewOdsSliderDark() = PreviewOdsSlider()
-
+@UiModePreviews.Default
 @Composable
 private fun PreviewOdsSliderLockups() = Preview {
-    val sliderValue = remember { mutableStateOf(50.0f) }
+    var value by remember { mutableStateOf(50.0f) }
     OdsSliderLockups(
-        value = sliderValue.value,
+        value = value,
         valueRange = 0f..100f,
-        onValueChange = { sliderValue.value = it }
+        onValueChange = { value = it }
     )
 }
-
-@Preview(name = "OdsSliderLockups - Light")
-@Composable
-private fun PreviewOdsSliderLockupsLight() = PreviewOdsSliderLockups()
-
-@Preview(
-    name = "OdsSliderLockups - Dark",
-    uiMode = Configuration.UI_MODE_NIGHT_YES,
-    showBackground = true
-)
-@Composable
-private fun PreviewOdsSliderLockupsDark() = PreviewOdsSliderLockups()