+++ lib/src/main/java/com/orange/ods/compose/component/control/OdsSwitch.kt
@@ -10,21 +10,22 @@
 
 package com.orange.ods.compose.component.control
 
-import android.content.res.Configuration
 import androidx.compose.material.ContentAlpha
 import androidx.compose.material.Switch
 import androidx.compose.material.SwitchDefaults
 import androidx.compose.material.ripple.LocalRippleTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.compositeOver
-import androidx.compose.ui.tooling.preview.Preview
 import com.orange.ods.compose.component.OdsComponentApi
 import com.orange.ods.compose.component.utilities.Preview
+import com.orange.ods.compose.component.utilities.UiModePreviews
 import com.orange.ods.compose.theme.OdsPrimaryRippleTheme
 import com.orange.ods.compose.theme.OdsTheme
 
@@ -92,24 +93,12 @@ private object OdsSwitchDefaults {
 
 }
 
-
+@UiModePreviews.Default
 @Composable
 private fun PreviewOdsSwitch() = Preview {
-    val checked = remember { mutableStateOf(false) }
+    var checked by remember { mutableStateOf(false) }
     OdsSwitch(
-        checked = checked.value,
-        onCheckedChange = { checked.value = it }
+        checked = checked,
+        onCheckedChange = { checked = it }
     )
 }
-
-@Preview(name = "OdsSwitch - Light")
-@Composable
-private fun PreviewOdsSwitchLight() = PreviewOdsSwitch()
-
-@Preview(
-    name = "OdsSwitch - Dark",
-    uiMode = Configuration.UI_MODE_NIGHT_YES,
-    showBackground = true
-)
-@Composable
-private fun PreviewOdsSwitchDark() = PreviewOdsSwitch()