+++ lib/src/main/java/com/orange/ods/compose/component/progressindicator/OdsLinearProgressIndicator.kt
@@ -13,7 +13,11 @@ package com.orange.ods.compose.component.progressindicator
 import androidx.compose.material.LinearProgressIndicator
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.PreviewParameter
 import com.orange.ods.compose.component.OdsComponentApi
+import com.orange.ods.compose.component.utilities.BasicPreviewParameterProvider
+import com.orange.ods.compose.component.utilities.Preview
+import com.orange.ods.compose.component.utilities.UiModePreviews
 import com.orange.ods.compose.theme.OdsTheme
 import com.orange.ods.utilities.extension.orElse
 
@@ -40,3 +44,11 @@ fun OdsLinearProgressIndicator(
         LinearProgressIndicator(modifier = modifier, color = progressIndicatorColor)
     }
 }
+
+@UiModePreviews.Default
+@Composable
+fun PreviewOdsLinearProgressIndicator(@PreviewParameter(OdsLinearProgressIndicatorPreviewParameterProvider::class) progress: Float?) = Preview {
+    OdsLinearProgressIndicator(progress = progress)
+}
+
+internal class OdsLinearProgressIndicatorPreviewParameterProvider : BasicPreviewParameterProvider<Float?>(0.75f, null)