+++ lib/src/main/java/com/orange/ods/compose/component/tab/OdsScrollableTabRow.kt
@@ -10,20 +10,21 @@
 
 package com.orange.ods.compose.component.tab
 
-import android.content.res.Configuration
 import androidx.annotation.DrawableRes
 import androidx.compose.material.ScrollableTabRow
 import androidx.compose.material.Tab
 import androidx.compose.material.TabRowDefaults
 import androidx.compose.material.TabRowDefaults.tabIndicatorOffset
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.tooling.preview.Preview
 import com.orange.ods.compose.component.OdsComponentApi
 import com.orange.ods.compose.component.utilities.Preview
+import com.orange.ods.compose.component.utilities.UiModePreviews
 import com.orange.ods.compose.theme.OdsTheme
 
 /**
@@ -63,38 +64,27 @@ fun OdsScrollableTabRow(
     )
 }
 
+@UiModePreviews.Default
 @Composable
 private fun PreviewOdsScrollableTabRow() = Preview {
     data class Tab(@DrawableRes val iconResId: Int, val text: String)
 
     val tabs = listOf(
-        Tab(android.R.drawable.ic_dialog_email, "First tab"),
-        Tab(android.R.drawable.ic_dialog_map, "Second tab"),
-        Tab(android.R.drawable.ic_dialog_dialer, "Third tab"),
-        Tab(android.R.drawable.ic_dialog_info, "Fourth tab")
+        Tab(android.R.drawable.ic_dialog_email, "First"),
+        Tab(android.R.drawable.ic_dialog_map, "Second"),
+        Tab(android.R.drawable.ic_dialog_dialer, "Third"),
+        Tab(android.R.drawable.ic_dialog_info, "Fourth")
     )
 
-    val selectedTabIndex = remember { mutableStateOf(0) }
-    OdsScrollableTabRow(selectedTabIndex = selectedTabIndex.value) {
+    var selectedTabIndex by remember { mutableStateOf(0) }
+    OdsScrollableTabRow(selectedTabIndex = selectedTabIndex) {
         tabs.forEachIndexed { index, tab ->
             OdsTab(
-                selected = selectedTabIndex.value == index,
-                onClick = { selectedTabIndex.value = index },
+                selected = selectedTabIndex == index,
+                onClick = { selectedTabIndex = index },
                 text = tab.text,
                 icon = painterResource(id = tab.iconResId)
             )
         }
     }
 }
-
-@Preview(name = "OdsScrollableTabRow - Light")
-@Composable
-private fun PreviewOdsScrollableTabRowLight() = PreviewOdsScrollableTabRow()
-
-@Preview(
-    name = "OdsScrollableTabRow - Dark",
-    uiMode = Configuration.UI_MODE_NIGHT_YES,
-    showBackground = true
-)
-@Composable
-private fun PreviewOdsScrollableTabRowDark() = PreviewOdsScrollableTabRow()