+++ lib/src/main/java/com/orange/ods/compose/component/textfield/OdsTextField.kt
@@ -10,7 +10,6 @@
 
 package com.orange.ods.compose.component.textfield
 
-import android.content.res.Configuration
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
@@ -38,13 +37,13 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.VisualTransformation
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import com.orange.ods.R
 import com.orange.ods.compose.component.OdsComponentApi
 import com.orange.ods.compose.component.utilities.BasicPreviewParameterProvider
 import com.orange.ods.compose.component.utilities.DisabledInteractionSource
 import com.orange.ods.compose.component.utilities.Preview
+import com.orange.ods.compose.component.utilities.UiModePreviews
 import com.orange.ods.compose.text.OdsTextCaption
 import com.orange.ods.compose.theme.OdsTheme
 import com.orange.ods.theme.OdsColors
@@ -232,38 +231,23 @@ internal object OdsTextFieldDefaults {
 
 }
 
+@UiModePreviews.Default
 @Composable
-private fun PreviewOdsTextField(hasCounter: Boolean) = Preview {
-    var text by remember { mutableStateOf("Input text") }
+private fun PreviewOdsTextField(@PreviewParameter(OdsTextFieldPreviewParameterProvider::class) hasCounter: Boolean) = Preview {
+    var value by remember { mutableStateOf("Input text") }
     Column {
         OdsTextField(
-            value = text,
-            onValueChange = { text = it },
+            value = value,
+            onValueChange = { value = it },
             placeholder = "Placeholder",
             leadingIcon = painterResource(id = android.R.drawable.ic_dialog_info),
             trailingIcon = painterResource(id = android.R.drawable.ic_input_add)
         )
 
         if (hasCounter) {
-            OdsTextFieldCounter(text.length, 30, Modifier.align(Alignment.End))
+            OdsTextFieldCounter(value.length, 30, Modifier.align(Alignment.End))
         }
     }
-}
-
-@Preview(name = "OdsTextField - Light")
-@Composable
-private fun PreviewOdsTextFieldLight(@PreviewParameter(OdsTextFieldPreviewParameterProvider::class) hasCounter: Boolean) {
-    PreviewOdsTextField(hasCounter)
-}
-
-@Preview(
-    name = "OdsTextField - Dark",
-    uiMode = Configuration.UI_MODE_NIGHT_YES,
-    showBackground = true
-)
-@Composable
-private fun PreviewOdsTextFieldDark(@PreviewParameter(OdsTextFieldPreviewParameterProvider::class) hasCounter: Boolean) {
-    PreviewOdsTextField(hasCounter)
 }
 
 internal class OdsTextFieldPreviewParameterProvider : BasicPreviewParameterProvider<Boolean>(false, true)