+++ lib/src/main/java/com/orange/ods/compose/component/textfield/password/OdsPasswordOutlinedTextField.kt
@@ -10,7 +10,6 @@
 
 package com.orange.ods.compose.component.textfield.password
 
-import android.content.res.Configuration
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.OutlinedTextField
@@ -25,10 +24,10 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.tooling.preview.Preview
 import com.orange.ods.compose.component.OdsComponentApi
 import com.orange.ods.compose.component.textfield.OdsOutlinedTextFieldDefaults
 import com.orange.ods.compose.component.utilities.Preview
+import com.orange.ods.compose.component.utilities.UiModePreviews
 import com.orange.ods.compose.theme.OdsTheme
 
 /**
@@ -99,24 +98,13 @@ fun OdsPasswordOutlinedTextField(
 
 }
 
+@UiModePreviews.Default
 @Composable
 private fun PreviewOdsPasswordOutlinedTextField() = Preview {
-    var text by remember { mutableStateOf("Input text") }
+    var value by remember { mutableStateOf("Input text") }
     OdsPasswordOutlinedTextField(
-        value = text,
-        onValueChange = { text = it },
+        value = value,
+        onValueChange = { value = it },
         placeholder = "Placeholder"
     )
 }
-
-@Preview(name = "OdsPasswordOutlinedTextField - Light")
-@Composable
-private fun PreviewOdsPasswordOutlinedTextFieldLight() = PreviewOdsPasswordOutlinedTextField()
-
-@Preview(
-    name = "OdsPasswordOutlinedTextField - Dark",
-    uiMode = Configuration.UI_MODE_NIGHT_YES,
-    showBackground = true
-)
-@Composable
-private fun PreviewOdsPasswordOutlinedTextFieldDark() = PreviewOdsPasswordOutlinedTextField()