+++ lib/src/main/java/com/orange/ods/compose/component/utilities/Preview.kt
@@ -10,8 +10,10 @@
 
 package com.orange.ods.compose.component.utilities
 
+import android.content.res.Configuration
 import androidx.compose.material.Surface
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import com.orange.ods.BuildConfig
 import com.orange.ods.compose.theme.OdsTheme
@@ -44,3 +46,37 @@ internal open class BasicPreviewParameterProvider<T>(vararg values: T) : Preview
  * @param clazz The enum class.
  */
 internal open class EnumPreviewParameterProvider(clazz: KClass<out Enum<*>>) : BasicPreviewParameterProvider<Enum<*>>(*clazz.java.enumConstants)
+
+/**
+ * Multipreview annotation classes used to display both light and dark mode previews.
+ *
+ * The only reason why `UiModePreviews` is an annotation class is to colorize it as an annotation in Android Studio.
+ * An empty `Target` annotation has been added in order to avoid using the parent `UiModePreviews` annotation which has no effect.
+ */
+@Target
+internal annotation class UiModePreviews {
+
+    companion object {
+        private const val LightName = "Light"
+        private const val DarkName = "Dark"
+        private const val ButtonWidthDp = 200
+        private const val ChipWidthDp = 180
+        private const val TabWidthDp = 100
+    }
+
+    @Preview(name = LightName)
+    @Preview(name = DarkName, uiMode = Configuration.UI_MODE_NIGHT_YES, showBackground = true)
+    annotation class Default
+
+    @Preview(name = LightName, widthDp = ButtonWidthDp)
+    @Preview(name = DarkName, uiMode = Configuration.UI_MODE_NIGHT_YES, showBackground = true, widthDp = ButtonWidthDp)
+    annotation class Button
+
+    @Preview(name = LightName, widthDp = ChipWidthDp)
+    @Preview(name = DarkName, uiMode = Configuration.UI_MODE_NIGHT_YES, showBackground = true, widthDp = ChipWidthDp)
+    annotation class Chip
+
+    @Preview(name = LightName, widthDp = TabWidthDp)
+    @Preview(name = DarkName, uiMode = Configuration.UI_MODE_NIGHT_YES, showBackground = true, widthDp = TabWidthDp)
+    annotation class Tab
+}